pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.2",
    "certifi>=2025.11.12",
    "fastapi>=0.124.4",
    "jinja2>=3.1.6",
    "livekit-agents>=1.3.6",
    "livekit-plugins-deepgram>=1.3.6",
    "livekit-plugins-openai>=1.3.6",
    "livekit-plugins-soniox>=1.3.6",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.9",
    "ty>=0.0.2",
]
